Optionally omit generating directives when instrumenting
This is used the python side so that we don't need to remove
the directives manually after instrumentation

+++ src/ValueRangeInstrumenter.h
@@ -21,7 +21,7 @@ private:
   std::map<std::string, clang::tooling::Replacements> &FileToReplacements;
   std::vector<RuleActionEditCollector> Rules;
   std::vector<clang::tooling::Replacement> Replacements;
-  std::map<std::string, int> FileToNumberMarkerDecls;
+  std::map<std::string, size_t> FileToNumberMarkerDecls;
 };
 
 } // namespace markers
